# app/static/swagger.yaml
swagger: "2.0"
info:
 title: "Auto Shop API"
 description: "API for managing an auto repair shop's mechanics, service tickets, and inventory"
 version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
 - "http"
consumes:
 - "application/json"
produces:
 - "application/json"
...
securityDefinitions:
 bearerAuth:
  type: apiKey
  name: Authorization
  in: header
  description: "JWT Authentication. Example: 'Bearer {token}'"
definitions:
...
 MechanicBase:
  type: "object"
  properties:
   name:
    type: "string"
    example: "John Smith"
   email:
    type: "string"
    format: "email"
    example: "john.smith@autoshop.com"
   phone:
    type: "string"
    example: "123-456-7890"
   salary:
    type: "number"
    format: "float"
    example: 75000.00
  required:
   - name
   - email
   - phone
   - salary
 InventoryBase:
  type: "object"
  properties:
   name:
    type: "string"
    example: "Oil Filter"
   price:
    type: "number"
    format: "float"
    minimum: 0
    example: 15.99
  required:
   - name
   - price
 InventoryResponse:
  type: "object"
  properties:
   id:
    type: "integer"
    example: 1
   name:
    type: "string"
    example: "Oil Filter"
   price:
    type: "number"
    format: "float"
    example: 15.99
   service_tickets:
    type: "array"
    items:
     type: "object"
     properties:
      id:
       type: "integer"
       example: 1
      VIN:
       type: "string"
       example: "1HGCM82633A123456"
      description:
       type: "string"
       example: "Oil change"
paths:
...
 /mechanics:
  post:
   tags:
    - "Mechanics"
   summary: "Create new mechanic"
   description: "Add a new mechanic to the system. Email must be unique."
   security:
    - bearerAuth: []
   parameters:
    - in: "body"
      name: "mechanic"
      description: "Mechanic information"
      required: true
      schema:
       $ref: "#/definitions/MechanicBase"
   responses:
    201:
     description: "Mechanic created successfully"
     schema:
      $ref: "#/definitions/MechanicBase"
    400:
     description: "Invalid input or duplicate email"
     schema:
      type: "object"
      properties:
       message:
        type: "string"
        example: "Email already registered"
       error:
        type: "string"
        example: "duplicate_email"
    401:
     description: "Unauthorized - Invalid or missing token"

  get:
   tags:
    - "Mechanics"
   summary: "Get all mechanics"
   description: "Get paginated list of mechanics"
   parameters:
    - in: "query"
      name: "page"
      type: "integer"
      description: "Page number"
      default: 1
    - in: "query"
      name: "per_page"
      type: "integer"
      description: "Items per page"
      default: 10
   responses:
    200:
     description: "Successfully retrieved mechanics"
     schema:
      type: "object"
      properties:
       mechanics:
        type: "array"
        items:
         $ref: "#/definitions/MechanicBase"
       total:
        type: "integer"
        example: 50
       pages:
        type: "integer"
        example: 5
       current_page:
        type: "integer"
        example: 1
       per_page:
        type: "integer"
        example: 10
       has_next:
        type: "boolean"
        example: true
       has_prev:
        type: "boolean"
        example: false
 /mechanics/{id}:
  parameters:
   - in: path
     name: id
     type: integer
     required: true
     description: "Mechanic ID"

  get:
   tags:
    - "Mechanics"
   summary: "Get mechanic by ID"
   description: "Retrieve a specific mechanic's information"
   responses:
    200:
     description: "Successfully retrieved mechanic"
     schema:
      $ref: "#/definitions/MechanicBase"
    404:
     description: "Mechanic not found"

  put:
   tags:
    - "Mechanics"
   summary: "Update mechanic"
   description: "Update mechanic information"
   security:
    - bearerAuth: []
   parameters:
    - in: "body"
      name: "mechanic"
      description: "Updated mechanic information"
      required: true
      schema:
       $ref: "#/definitions/MechanicBase"
   responses:
    200:
     description: "Mechanic updated successfully"
     schema:
      $ref: "#/definitions/MechanicBase"
    404:
     description: "Mechanic not found"
    401:
     description: "Unauthorized - Invalid or missing token"

  delete:
   tags:
    - "Mechanics"
   summary: "Delete mechanic"
   description: "Delete a mechanic from the system"
   security:
    - bearerAuth: []
   responses:
    200:
     description: "Mechanic deleted successfully"
     schema:
      type: "object"
      properties:
       message:
        type: "string"
        example: "Mechanic deleted successfully"
    404:
     description: "Mechanic not found"
    401:
     description: "Unauthorized - Invalid or missing token"
 /mechanics/tickets:
  get:
   tags:
    - "Mechanics"
   summary: "Get mechanic's assigned tickets"
   description: "Get all service tickets assigned to the authenticated mechanic"
   security:
    - bearerAuth: []
   responses:
    200:
     description: "Successfully retrieved tickets"
     schema:
      type: "object"
      properties:
       tickets:
        type: "array"
        items:
         type: "object"
         properties:
          id:
           type: "integer"
           example: 1
          VIN:
           type: "string"
           example: "1HGCM82633A123456"
          description:
           type: "string"
           example: "Oil change and tire rotation"
          service_date:
           type: "string"
           format: "date"
           example: "2024-02-15"
    401:
     description: "Unauthorized - Invalid or missing token"
    404:
     description: "Mechanic not found"
 /inventory:
  post:
   tags:
    - "Inventory"
   summary: "Create new part"
   description: "Add a new part to inventory"
   security:
    - bearerAuth: []
   parameters:
    - in: "body"
      name: "part"
      description: "Part information"
      required: true
      schema:
       $ref: "#/definitions/InventoryBase"
   responses:
    201:
     description: "Part created successfully"
     schema:
      $ref: "#/definitions/InventoryResponse"
    400:
     description: "Invalid input"
     schema:
      type: "object"
      properties:
       message:
        type: "string"
        example: "Database error occurred"
       error:
        type: "string"
        example: "integrity_error"
    401:
     description: "Unauthorized - Invalid or missing token"

  get:
   tags:
    - "Inventory"
   summary: "Get all parts"
   description: "Get paginated list of inventory parts"
   parameters:
    - in: "query"
      name: "page"
      type: "integer"
      description: "Page number"
      default: 1
    - in: "query"
      name: "per_page"
      type: "integer"
      description: "Items per page"
      default: 10
   responses:
    200:
     description: "Successfully retrieved parts"
     schema:
      type: "object"
      properties:
       parts:
        type: "array"
        items:
         $ref: "#/definitions/InventoryResponse"
       total:
        type: "integer"
        example: 50
       pages:
        type: "integer"
        example: 5
       current_page:
        type: "integer"
        example: 1
       per_page:
        type: "integer"
        example: 10
       has_next:
        type: "boolean"
        example: true
       has_prev:
        type: "boolean"
        example: false
 /inventory/{part_id}:
  parameters:
   - in: path
     name: part_id
     type: integer
     required: true
     description: "Part ID"

  put:
   tags:
    - "Inventory"
   summary: "Update part"
   description: "Update part information"
   security:
    - bearerAuth: []
   parameters:
    - in: "body"
      name: "part"
      description: "Updated part information"
      required: true
      schema:
       $ref: "#/definitions/InventoryBase"
   responses:
    200:
     description: "Part updated successfully"
     schema:
      $ref: "#/definitions/InventoryResponse"
    400:
     description: "Invalid input"
     schema:
      type: "object"
      properties:
       message:
        type: "string"
        example: "Database error occurred"
    401:
     description: "Unauthorized - Invalid or missing token"
    404:
     description: "Part not found"

  delete:
   tags:
    - "Inventory"
   summary: "Delete part"
   description: "Delete a part from inventory"
   security:
    - bearerAuth: []
   responses:
    200:
     description: "Part deleted successfully"
     schema:
      type: "object"
      properties:
       message:
        type: "string"
        example: "Part deleted successfully"
    400:
     description: "Part in use"
     schema:
      type: "object"
      properties:
       message:
        type: "string"
        example: "Cannot delete part that is used in service tickets"
       ticket_count:
        type: "integer"
        example: 2
    401:
     description: "Unauthorized - Invalid or missing token"
    404:
     description: "Part not found"
...